Switcheroo
Moved Main stuff to EventStorage and editted save/load from a file
